Experimenting with using semantic LSP highlighting
Finally getting around to trying this out with the new tokyonight theme.
It gives me a bit more fidelity and also allows me to target things a
bit more.  Something I definitely miss from my old Evokai was a bit mroe
separation on stuff. On really gnarly area I've noticed for myself
recently is that the comment style brackets has really fucked with my
ability to parse out what I am looking at.  But I've gone in now and
added a bit more fidelity on the theme and it's definitely becoming a
bit easier to parse